+++ Sources/MapboxMaps/SwiftUI/Builders/MapContentVisitor.swift
@@ -1,19 +1,12 @@
 import SwiftUI
 
-/// Type erasure wrapper for AnnotationsGroup.
-struct AnyAnnotationGroup {
-    var layerId: String?
-    var update: (AnnotationOrchestrator, String, inout [AnyHashable: String]) -> Void
-}
-
 protocol MapContentVisitor: AnyObject {
     var id: [AnyHashable] { get }
-    func push(_ idPart: AnyHashable)
-    func pop()
+    var locationOptions: LocationOptions { get set }
     @available(iOS 13.0, *)
     func add(viewAnnotation: MapViewAnnotation)
-    var locationOptions: LocationOptions { get set }
-    func add(annotationGroup: AnyAnnotationGroup)
+    func add(annotationGroup: AnnotationGroup)
+    func visit(id: AnyHashable, content: MapContent)
 }
 
 @available(iOS 13.0, *)
@@ -23,22 +16,20 @@ final class DefaultMapContentVisitor: MapContentVisitor {
     private(set) var id: [AnyHashable] = []
     private(set) var visitedViewAnnotations: [AnyHashable: MapViewAnnotation] = [:]
 
-    private(set) var annotationGroups: [(AnyHashable, AnyAnnotationGroup)] = []
-
-    func push(_ idPart: AnyHashable) {
-        id.append(idPart)
-    }
-
-    func pop() {
-        id.removeLast()
-    }
+    private(set) var annotationGroups: [(AnyHashable, AnnotationGroup)] = []
 
     @available(iOS 13.0, *)
     func add(viewAnnotation: MapViewAnnotation) {
         visitedViewAnnotations[id] = viewAnnotation
     }
 
-    func add(annotationGroup: AnyAnnotationGroup) {
+    func add(annotationGroup: AnnotationGroup) {
         annotationGroups.append((id, annotationGroup))
+    }
+
+    func visit(id: AnyHashable, content: MapContent) {
+        self.id.append(id)
+        content.visit(self)
+        self.id.removeLast()
     }
 }